+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Redactor for Craft CMS
 
+## Unreleased
+
+- Fixed several styling issues with Redactor’s fullscreen mode. ([#450](https://github.com/craftcms/redactor/pull/450), [#445](https://github.com/craftcms/redactor/issues/445))
+
 ## 2.10.11 - 2022-12-31
 
 - Fixed a bug where image URLs would be broken after selecting them, if the URL contained any underscores. ([#441](https://github.com/craftcms/redactor/issues/441))